Refactor ui package
