+++ src/main/kotlin/ui/ConfigurationEditor.kt
@@ -1,81 +1,43 @@
 package ui
 
-import csstype.AlignSelf
-import csstype.Display
-import csstype.Length
-import csstype.Margin
-import csstype.VerticalAlign
-import csstype.em
 import csstype.px
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import kotlinx.js.jso
 import model.ExportResult
 import model.Feature
 import model.Format
 import model.LyricsType
-import model.LyricsType.KanaCv
-import model.LyricsType.KanaVcv
-import model.LyricsType.RomajiCv
-import model.LyricsType.RomajiVcv
 import model.LyricsType.Unknown
 import model.Project
 import model.TICKS_IN_FULL_NOTE
-import mui.icons.material.AddCircle
-import mui.icons.material.ArrowDownward
-import mui.icons.material.ArrowUpward
-import mui.icons.material.ErrorOutline
-import mui.icons.material.HelpOutline
-import mui.icons.material.RemoveCircle
-import mui.material.BaseTextFieldProps
-import mui.material.Box
 import mui.material.Button
 import mui.material.ButtonColor
 import mui.material.ButtonVariant
-import mui.material.FormControl
-import mui.material.FormControlLabel
-import mui.material.FormControlMargin
-import mui.material.FormControlVariant
-import mui.material.FormGroup
-import mui.material.FormLabel
-import mui.material.IconButton
-import mui.material.IconButtonColor
-import mui.material.LabelPlacement
-import mui.material.MenuItem
-import mui.material.Paper
-import mui.material.Radio
-import mui.material.RadioColor
-import mui.material.RadioGroup
-import mui.material.StandardTextFieldProps
-import mui.material.Switch
-import mui.material.SwitchColor
-import mui.material.TextField
-import mui.material.Tooltip
-import mui.material.TooltipPlacement
-import mui.material.Typography
-import mui.material.styles.TypographyVariant
-import mui.system.sx
 import process.RESTS_FILLING_MAX_LENGTH_DENOMINATOR_DEFAULT
 import process.evalFractionOrNull
 import process.fillRests
 import process.lyrics.LyricsReplacementRequest
 import process.lyrics.convert
 import process.lyrics.replaceLyrics
-import process.needWarningZoom
 import process.projectZoomFactorOptions
-import process.restsFillingMaxLengthDenominatorOptions
 import process.zoom
 import react.ChildrenBuilder
-import react.ElementType
-import react.FC
 import react.Props
-import react.ReactNode
-import react.StateSetter
-import react.create
 import react.css.css
 import react.dom.html.ReactHTML.div
 import react.useState
+import ui.common.DialogErrorState
+import ui.common.SubState
+import ui.common.errorDialog
+import ui.common.progress
+import ui.common.scopedFC
+import ui.common.title
+import ui.configuration.LyricsConversionBlock
+import ui.configuration.LyricsReplacementBlock
+import ui.configuration.PitchConversionBlock
+import ui.configuration.ProjectZoomBlock
+import ui.configuration.SlightRestsFillingBlock
 import ui.strings.Strings
 import ui.strings.string
 import util.runCatchingCancellable
@@ -184,590 +146,6 @@ val ConfigurationEditor = scopedFC<ConfigurationEditorProps> { props, scope ->
     )
 
     progress(isShowing = isProcessing)
-}
-
-external interface LyricsConversionProps : SubProps<LyricsConversionState> {
-    var project: Project
-    var outputFormat: Format
-}
-
-private val LyricsConversionBlock = subFC<LyricsConversionProps, LyricsConversionState> { props, state, editState ->
-    FormGroup {
-        buildFeatureSwitch(
-            isOn = state.isOn,
-            onSwitched = { editState { copy(isOn = it) } },
-            labelStrings = Strings.JapaneseLyricsConversion
-        )
-    }
-
-    if (state.isOn) buildLyricsDetail(
-        props = props,
-        fromLyricsType = state.fromType,
-        setFromLyricsType = { editState { copy(fromType = it) } },
-        toLyricsType = state.toType,
-        setToLyricsType = { editState { copy(toType = it) } },
-    )
-}
-
-private fun ChildrenBuilder.buildLyricsDetail(
-    props: LyricsConversionProps,
-    fromLyricsType: LyricsType?,
-    setFromLyricsType: (LyricsType) -> Unit,
-    toLyricsType: LyricsType?,
-    setToLyricsType: (LyricsType) -> Unit
-) {
-    div {
-        css {
-            margin = Margin(horizontal = 40.px, vertical = 0.px)
-            width = Length.maxContent
-        }
-        Paper {
-            elevation = 0
-            div {
-                css {
-                    margin = Margin(
-                        horizontal = 24.px,
-                        top = 16.px,
-                        bottom = 24.px
-                    )
-                }
-                FormGroup {
-                    buildLyricsTypeControl(
-                        labelText = string(
-                            Strings.FromLyricsTypeLabel,
-                            "type" to props.project.lyricsType.displayName
-                        ),
-                        type = fromLyricsType,
-                        setType = setFromLyricsType,
-                        lyricTypeOptions = listOf(RomajiCv, RomajiVcv, KanaCv, KanaVcv)
-                    )
-
-                    buildLyricsTypeControl(
-                        labelText = string(Strings.ToLyricsTypeLabel),
-                        type = toLyricsType,
-                        setType = setToLyricsType,
-                        lyricTypeOptions = props.outputFormat.possibleLyricsTypes
-                    )
-                }
-            }
-        }
-    }
-}
-
-private fun ChildrenBuilder.buildLyricsTypeControl(
-    labelText: String,
-    type: LyricsType?,
-    setType: (LyricsType) -> Unit,
-    lyricTypeOptions: List<LyricsType>
-) {
-    FormControl {
-        margin = FormControlMargin.normal
-        FormLabel {
-            focused = false
-            +labelText
-        }
-        RadioGroup {
-            row = true
-            value = type?.name.orEmpty()
-            onChange = { event, _ ->
-                val value = event.target.value
-                setType(LyricsType.valueOf(value))
-            }
-            lyricTypeOptions.forEach { lyricsType ->
-                FormControlLabel {
-                    value = lyricsType.name
-                    control = Radio.create {
-                        color = RadioColor.secondary
-                    }
-                    label = Typography.create {
-                        variant = TypographyVariant.subtitle2
-                        +lyricsType.displayName
-                    }
-                }
-            }
-        }
-    }
-}
-
-external interface LyricsReplacementProps : SubProps<LyricsReplacementState>
-
-private val LyricsReplacementBlock = subFC<LyricsReplacementProps, LyricsReplacementState> { _, state, editState ->
-    FormGroup {
-        buildFeatureSwitch(
-            isOn = state.isOn,
-            onSwitched = { editState { copy(isOn = it) } },
-            labelStrings = Strings.LyricsReplacement
-        )
-    }
-
-    if (state.isOn) buildLyricsReplacementDetail(state, editState)
-}
-
-private fun ChildrenBuilder.buildLyricsReplacementDetail(
-    state: LyricsReplacementState,
-    editState: (LyricsReplacementState.() -> LyricsReplacementState) -> Unit
-) {
-    div {
-        css {
-            margin = Margin(horizontal = 40.px, vertical = 0.px)
-            width = Length.maxContent
-        }
-        Paper {
-            elevation = 0
-            div {
-                css {
-                    margin = Margin(
-                        horizontal = 24.px,
-                        top = 16.px,
-                        bottom = 24.px
-                    )
-                    paddingTop = 8.px
-                    paddingBottom = 8.px
-                }
-                state.request.items.forEachIndexed { index, item ->
-                    buildLyricsReplacementItem(index, item, editState)
-                }
-                div {
-                    Button {
-                        color = ButtonColor.secondary
-                        variant = ButtonVariant.text
-                        AddCircle()
-                        onClick = { editState { copy(request = request.add()) } }
-                        div {
-                            css { padding = 8.px }
-                            +string(Strings.LyricsReplacementAddItemButton)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun ChildrenBuilder.buildLyricsReplacementItem(
-    index: Int,
-    item: LyricsReplacementRequest.Item,
-    editState: (LyricsReplacementState.() -> LyricsReplacementState) -> Unit
-) {
-    fun editRequest(block: LyricsReplacementRequest.() -> LyricsReplacementRequest) {
-        editState { copy(request = request.block()) }
-    }
-
-    fun editItem(block: LyricsReplacementRequest.Item.() -> LyricsReplacementRequest.Item) {
-        editRequest { copy(items = items.mapIndexed { i, it -> if (i == index) block(it) else it }) }
-    }
-    div {
-        css {
-            display = Display.flex
-            marginBottom = 16.px
-        }
-        Typography {
-            css {
-                color = appTheme.palette.secondary.main
-                alignSelf = AlignSelf.center
-            }
-            variant = TypographyVariant.subtitle2
-            component = "span".asDynamic().unsafeCast<ElementType<*>>()
-            +string(Strings.LyricsReplacementItemLabel, "number" to (index + 1).toString())
-        }
-        FormControl {
-            style = jso {
-                marginLeft = 2.em
-                marginTop = 8.px
-                marginBottom = 8.px
-            }
-            variant = FormControlVariant.standard
-            focused = false
-            FormLabel {
-                focused = false
-                Typography {
-                    variant = TypographyVariant.caption
-                    +string(Strings.LyricsReplacementFilterTypeLabel)
-                }
-            }
-            TextField {
-                sx { minWidth = 5.em }
-                select = true
-                value = item.filterType.unsafeCast<Nothing?>()
-                (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                    val value = event.target.asDynamic().value as String
-                    editItem { copy(filterType = LyricsReplacementRequest.FilterType.valueOf(value)) }
-                }
-                LyricsReplacementRequest.FilterType.values().forEach { type ->
-                    MenuItem {
-                        value = type.toString()
-                        +string(type.strings)
-                    }
-                }
-            }
-        }
-        FormControl {
-            style = jso {
-                marginLeft = 2.em
-                marginTop = 8.px
-                marginBottom = 8.px
-            }
-
-            focused = false
-            FormLabel {
-                focused = false
-                Typography {
-                    variant = TypographyVariant.caption
-                    +string(Strings.LyricsReplacementFilterTextLabel)
-                }
-            }
-            TextField {
-                sx { width = 8.em }
-                value = item.filter
-                disabled = item.filterType.needsFilter().not()
-                (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                    val value = event.target.asDynamic().value as String
-                    editItem { copy(filter = value) }
-                }
-            }
-        }
-        FormControl {
-            style = jso {
-                marginLeft = 2.em
-                marginTop = 8.px
-                marginBottom = 8.px
-            }
-            variant = FormControlVariant.standard
-            focused = false
-            FormLabel {
-                focused = false
-                Typography {
-                    variant = TypographyVariant.caption
-                    +string(Strings.LyricsReplacementMatchTypeLabel)
-                }
-            }
-            TextField {
-                sx { minWidth = 5.em }
-                select = true
-                value = item.matchType.unsafeCast<Nothing?>()
-                (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                    val value = event.target.asDynamic().value as String
-                    editItem { copy(matchType = LyricsReplacementRequest.MatchType.valueOf(value)) }
-                }
-                LyricsReplacementRequest.MatchType.values().forEach { type ->
-                    MenuItem {
-                        value = type.toString()
-                        +string(type.strings)
-                    }
-                }
-            }
-        }
-        FormControl {
-            style = jso {
-                marginLeft = 2.em
-                marginTop = 8.px
-                marginBottom = 8.px
-            }
-            focused = false
-            FormLabel {
-                focused = false
-                Typography {
-                    variant = TypographyVariant.caption
-                    +string(Strings.LyricsReplacementFromTextLabel)
-                }
-            }
-            TextField {
-                sx { width = 8.em }
-                value = item.from
-                disabled = item.matchType.needsFrom().not()
-                (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                    val value = event.target.asDynamic().value as String
-                    editItem { copy(from = value) }
-                }
-            }
-        }
-        FormControl {
-            style = jso {
-                marginLeft = 2.em
-                marginTop = 8.px
-                marginBottom = 8.px
-            }
-            focused = false
-            FormLabel {
-                focused = false
-                Typography {
-                    variant = TypographyVariant.caption
-                    +string(Strings.LyricsReplacementToTextLabel)
-                }
-            }
-            TextField {
-                sx { width = 8.em }
-                value = item.to
-                (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                    val value = event.target.asDynamic().value as String
-                    editItem { copy(to = value) }
-                }
-            }
-        }
-        IconButton {
-            color = IconButtonColor.inherit
-            style = jso {
-                margin = 5.px
-                marginLeft = 20.px
-                height = Length.fitContent
-                alignSelf = AlignSelf.center
-            }
-            onClick = {
-                editRequest { moveUp(index) }
-            }
-            ArrowUpward()
-        }
-        IconButton {
-            color = IconButtonColor.inherit
-            style = jso {
-                margin = 5.px
-                height = Length.fitContent
-                alignSelf = AlignSelf.center
-            }
-            onClick = {
-                editRequest { moveDown(index) }
-            }
-            ArrowDownward()
-        }
-        IconButton {
-            color = IconButtonColor.secondary
-            style = jso {
-                margin = 5.px
-                height = Length.fitContent
-                alignSelf = AlignSelf.center
-            }
-            onClick = {
-                editRequest { remove(index) }
-            }
-            RemoveCircle()
-        }
-    }
-}
-
-external interface SlightRestsFillingProps : SubProps<SlightRestsFillingState>
-
-private val SlightRestsFillingBlock = subFC<SlightRestsFillingProps, SlightRestsFillingState> { _, state, editState ->
-    FormGroup {
-        div {
-            buildFeatureSwitch(
-                isOn = state.isOn,
-                onSwitched = { editState { copy(isOn = it) } },
-                labelStrings = Strings.SlightRestsFilling
-            )
-            Tooltip {
-                title = ReactNode(string(Strings.SlightRestsFillingDescription))
-                placement = TooltipPlacement.right
-                disableInteractive = false
-                HelpOutline {
-                    style = jso {
-                        verticalAlign = VerticalAlign.middle
-                    }
-                }
-            }
-        }
-    }
-
-    if (state.isOn) buildRestsFillingDetail(state, editState)
-}
-
-private fun ChildrenBuilder.buildRestsFillingDetail(
-    state: SlightRestsFillingState,
-    editState: (SlightRestsFillingState.() -> SlightRestsFillingState) -> Unit
-) {
-    div {
-        css {
-            margin = Margin(horizontal = 40.px, vertical = 0.px)
-            width = Length.maxContent
-        }
-        Paper {
-            elevation = 0
-            Box {
-                style = jso {
-                    margin = Margin(
-                        left = 24.px,
-                        right = 48.px,
-                        top = 16.px,
-                        bottom = 16.px
-                    )
-                    paddingBottom = 8.px
-                }
-                sx { minWidth = 15.em }
-                FormControl {
-                    margin = FormControlMargin.normal
-                    variant = FormControlVariant.standard
-                    focused = false
-                    FormLabel {
-                        focused = false
-                        Typography {
-                            variant = TypographyVariant.caption
-                            +string(Strings.SlightRestsFillingThresholdLabel)
-                        }
-                    }
-                    TextField {
-                        style = jso { minWidth = 5.em }
-                        select = true
-                        value = state.excludedMaxLengthDenominator.toString().unsafeCast<Nothing?>()
-                        (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                        (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                            val value = event.target.asDynamic().value as String
-                            editState { copy(excludedMaxLengthDenominator = value.toInt()) }
-                        }
-                        restsFillingMaxLengthDenominatorOptions.forEach { denominator ->
-                            MenuItem {
-                                value = denominator.toString()
-                                +string(
-                                    Strings.SlightRestsFillingThresholdItem,
-                                    "denominator" to denominator.toString()
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-external interface PitchConversionProps : SubProps<PitchConversionState>
-
-private val PitchConversionBlock = subFC<PitchConversionProps, PitchConversionState> { _, state, editState ->
-    FormGroup {
-        div {
-            buildFeatureSwitch(
-                isOn = state.isOn,
-                onSwitched = { editState { copy(isOn = it) } },
-                labelStrings = Strings.ConvertPitchData
-            )
-            Tooltip {
-                title = ReactNode(string(Strings.ConvertPitchDataDescription))
-                placement = TooltipPlacement.right
-                disableInteractive = false
-                ErrorOutline {
-                    style = jso {
-                        verticalAlign = VerticalAlign.middle
-                    }
-                }
-            }
-        }
-    }
-}
-
-external interface ProjectZoomProps : SubProps<ProjectZoomState> {
-    var project: Project
-}
-
-private val ProjectZoomBlock = subFC<ProjectZoomProps, ProjectZoomState> { props, state, editState ->
-    FormGroup {
-        div {
-            buildFeatureSwitch(
-                isOn = state.isOn,
-                onSwitched = { editState { copy(isOn = it) } },
-                labelStrings = Strings.ProjectZoom
-            )
-            Tooltip {
-                title = ReactNode(string(Strings.ProjectZoomDescription))
-                placement = TooltipPlacement.right
-                disableInteractive = false
-                HelpOutline {
-                    style = jso {
-                        verticalAlign = VerticalAlign.middle
-                    }
-                }
-            }
-            if (props.project.needWarningZoom(state.factorValue)) {
-                Tooltip {
-                    title = ReactNode(string(Strings.ProjectZoomWarning))
-                    placement = TooltipPlacement.right
-                    disableInteractive = false
-                    ErrorOutline {
-                        style = jso {
-                            verticalAlign = VerticalAlign.middle
-                        }
-                    }
-                }
-            }
-        }
-    }
-    if (state.isOn) buildProjectZoomDetail(state, editState)
-}
-
-private fun ChildrenBuilder.buildProjectZoomDetail(
-    state: ProjectZoomState,
-    editState: (ProjectZoomState.() -> ProjectZoomState) -> Unit
-) {
-    div {
-        css {
-            margin = Margin(horizontal = 40.px, vertical = 0.px)
-            width = Length.maxContent
-        }
-        Paper {
-            elevation = 0
-            Box {
-                style = jso {
-                    margin = Margin(
-                        left = 24.px,
-                        right = 48.px,
-                        top = 16.px,
-                        bottom = 16.px
-                    )
-                    paddingBottom = 8.px
-                }
-                sx { minWidth = 15.em }
-                FormControl {
-                    FormLabel {
-                        focused = false
-                        Typography {
-                            variant = TypographyVariant.caption
-                            +string(Strings.ProjectZooLabel)
-                        }
-                    }
-                    margin = FormControlMargin.normal
-                    variant = FormControlVariant.standard
-                    focused = false
-                    TextField {
-                        style = jso { minWidth = 5.em }
-                        select = true
-                        value = state.factor.unsafeCast<Nothing?>()
-                        (this.unsafeCast<BaseTextFieldProps>()).variant = FormControlVariant.standard
-                        (this.unsafeCast<StandardTextFieldProps>()).onChange = { event ->
-                            val value = event.target.asDynamic().value as String
-                            editState { copy(factor = value) }
-                        }
-                        projectZoomFactorOptions.forEach { factor ->
-                            MenuItem {
-                                value = factor
-                                +(factor)
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun ChildrenBuilder.buildFeatureSwitch(
-    isOn: Boolean,
-    onSwitched: (Boolean) -> Unit,
-    labelStrings: Strings
-) {
-    FormControlLabel {
-        label = ReactNode(string(labelStrings))
-        control = Switch.create {
-            color = SwitchColor.secondary
-            checked = isOn
-            onChange = { event, _ ->
-                val checked = event.target.checked
-                onSwitched(checked)
-            }
-        }
-        labelPlacement = LabelPlacement.end
-    }
 }
 
 private fun ChildrenBuilder.buildNextButton(
@@ -882,25 +260,6 @@ external interface ConfigurationEditorProps : Props {
     var project: Project
     var outputFormat: Format
     var onFinished: (ExportResult, Format) -> Unit
-}
-
-external interface SubProps<T : SubState> : Props {
-    var initialState: T
-    var submitState: StateSetter<T>
-}
-
-interface SubState
-
-fun <P : SubProps<T>, T : SubState> subFC(
-    block: ChildrenBuilder.(props: P, state: T, editState: (T.() -> T) -> Unit) -> Unit,
-) = FC<P> { props ->
-    var state by useState(props.initialState)
-    fun editState(editor: T.() -> T) {
-        val newState = editor(state)
-        state = newState
-        props.submitState(newState)
-    }
-    block(props, state, ::editState)
 }
 
 data class LyricsConversionState(
