+++ src/main/kotlin/ui/common/ErrorDialog.kt
@@ -1,4 +1,4 @@
-package ui
+package ui.common
 
 import csstype.px
 import kotlinx.browser.window
